[ias-verify] move netscape_comment to own module, move safe_indexing stuff to own utils
